Added auto converting to BigInteger from Integer for to-java since there is no literal syntax in Clojure for BigInteger

+++ src/test/java/offbytwo/beanhelpers/test/Person.java
@@ -1,40 +1,42 @@
 package offbytwo.beanhelpers.test;
 
+import java.math.BigInteger
+
 public class Person {
 	private String name;
-	private int age;
+	private BigInteger age;
 	private Address address;
 
     public Person() {
     }
 
-    public Person(String name, int age, Address address) {
+    public Person(String name, BigInteger age, Address address) {
         this.name = name;
         this.age = age;
         this.address = address;
     }
 
     public String getName() {
-		return name;
-	}
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	public int getAge() {
-		return age;
-	}
+    public BigInteger getAge() {
+        return age;
+    }
 
-	public void setAge(int age) {
-		this.age = age;
-	}
+    public void setAge(BigInteger age) {
+        this.age = age;
+    }
 
-	public Address getAddress() {
-		return address;
-	}
+    public Address getAddress() {
+        return address;
+    }
 
-	public void setAddress(Address address) {
-		this.address = address;
-	}
+    public void setAddress(Address address) {
+        this.address = address;
+    }
 }
